% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_preprocessing.R
\name{add_means}
\alias{add_means}
\title{Return a SMET_OBJECT with daily, monthly, seasonnal, and yerly means added}
\usage{
add_means(data)
}
\arguments{
\item{data}{SMET_OBJECT}
}
\value{
SMET_OBJECT
}
\description{
Return a SMET_OBJECT with daily, monthly, seasonnal and yearly means added.
The\code{\link[base]{data.frame}} added are \code{$daily} (daily means),
\code{$monthly} (monthly means), \code{$yearly} (yearly means),
\code{$DJF} (winter means), \code{$MAM} (spring means),
\code{$JJA} (summer means), and  \code{$SON} (fall means)
}
