% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STL_analysis.R
\name{add_meteo}
\alias{add_meteo}
\title{Add meteo data to SMET_OBJECT}
\usage{
add_meteo(data, meteo_data)
}
\arguments{
\item{data}{SMET_OBJECT containing river data}

\item{meteo_data}{SMET_OBJECT continning meteo data}
}
\value{
SMET_OBJECT  with meteo added as a SMET_OBJECT object list \code{$meteo_ts}
}
\description{
Read SMET_OBJECT , returns a SMET_OBJECT with meteo stations indicated in
\code{$meteo} added as a SMET_OBJECT object list in \code{$meteo_ts}.
}
