% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STL_analysis.R
\name{add_meteo_and_trim}
\alias{add_meteo_and_trim}
\title{Wrapper function}
\usage{
add_meteo_and_trim(data, meteo_data, start, end)
}
\arguments{
\item{data}{A SMET_OBJECT with river data}

\item{meteo_data}{A SMET_OBJECT with meteo data}

\item{start}{Starting year to keep}

\item{end}{Ending year to keep}
}
\value{
SMET_OBJECT cut to the given dates with meteo data added
}
\description{
Wrapper for \code{\link{cut_all_data}} and \code{\link{add_meteo}} functions
}
