% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hysteresis_analysis.R
\name{check_and_cut_variables}
\alias{check_and_cut_variables}
\title{Check that variable and cut them to the same length}
\usage{
check_and_cut_variables(data)
}
\arguments{
\item{data}{A SMET_OBJECT obtained through the function
\code{\link{get_file_data}} containing the data of one river station}
}
\value{
A SMET_OBJECT with only the \code{$data} part filled with the cut
daily means T and Q time series.
}
\description{
Check that both water temperature (T) and discharge (Q) are provided and
cut the daily means of the two variables to have the same start and end
}
