% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hysteresis_analysis.R
\name{compute_daily_means_over_period}
\alias{compute_daily_means_over_period}
\title{Compute the mean of Julian days value over a period}
\usage{
compute_daily_means_over_period(data, period)
}
\arguments{
\item{data}{A SMET_OBJECT obtained through the function
\code{\link{get_file_data}} containing the data of one river station}

\item{period}{An integer representing the number of year over whihch
the average should be computed}
}
\value{
A list with the period as key. Periods have the format "AAAA-BBBB" where
AAAA is the starting year and BBBB the ending year. Each list element contains
a \code{\link[base]{data.frame}} with columns T and Q, which are  the day-of-the-yaer mean over the period
for the water temperature and the discharge.
}
\description{
Compute the day-of-the-yaer mean over a period of a given length in year.
The day-of-the-yaer mean is the mean between each first of jannuary, each
second of jannuary, and-so-on. The periods are tkaen from the end of the available
data and as many periods as possible are taken.
}
