% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STL_analysis.R
\name{cut_all_data}
\alias{cut_all_data}
\title{Cut timeseries in a SMET_OBJECT}
\usage{
cut_all_data(data, start, end)
}
\arguments{
\item{data}{SMET_OBJECT}

\item{start}{Starting year to keep}

\item{end}{Ending year to keep}
}
\value{
SMET_OBJECT
}
\description{
Read a SMET_OBJECT, returns a SMET cut between the years indicated by start and end
for the inner \code{\link[base]{data.frame}} \code{$data},\code{$daily},
\code{$monthly}, and \code{$yearly}
}
