% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_preprocessing.R
\name{cut_data}
\alias{cut_data}
\title{Read a SMET_OBJECT object, returns a SMET cut between the years indicated by start and end}
\usage{
cut_data(data, start, end)
}
\arguments{
\item{data}{A SMET_OBJECT}

\item{start}{Starting year to keep}

\item{end}{Ending year to keep}
}
\value{
A SMET_OBJECT where timeseries are cut between start and end
}
\description{
Read a SMET_OBJECT object, returns a SMET cut between the years indicated by start and end
}
