% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_preprocessing.R
\name{cut_full_year}
\alias{cut_full_year}
\title{Cut SMET_OBJECT to only keep complete years}
\usage{
cut_full_year(data)
}
\arguments{
\item{data}{SMET_OBJECT}
}
\value{
SMET_OBJECT
}
\description{
Cut SMET_OBJECT to only keep complete years (partial years at the beggining
or at the end of the time series are removed)
}
