% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{daily_mean_for_year}
\alias{daily_mean_for_year}
\title{Compute the daily mean over a year}
\usage{
daily_mean_for_year(data, yr)
}
\arguments{
\item{data}{A SMET_OBJECT for a given station}

\item{yr}{The year over which the data should be daily averaged}
}
\value{
A \code{\link[base]{data.frame}} the daily mean of each variables
(the timestamp is removed).
}
\description{
Compute the daily mean over a year for all raw variables in SMET_OBJECT
obtained through the function \code{\link{get_file_data}} (raw variables are stored
in the \code{$data} part of theSMET_OBJECT).
}
