% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_functions.R
\name{genreal_analysis_plots}
\alias{genreal_analysis_plots}
\title{Plot variable distributions and compute t-tests}
\usage{
genreal_analysis_plots(period, rivers_data)
}
\arguments{
\item{period}{A string, either "1999-2018", "1979-1998", "1979-2018", or
"1970-2018" defining the period over which the plots and analysis should be
produced}

\item{rivers_data}{Either \code{"NONE"} (default), \code{"PDF"} or \code{"PNG"}.}
}
\description{
This function produces the plots  variable distributions as shown in Figures
5, 6, and 10 and in Figures S11 to S13 in supplementary. It also computes
wilcox test shown in Table 2 and print the results to the console.
}
