% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_preprocessing.R
\name{get_data}
\alias{get_data}
\title{Read raw of river or meteo data}
\usage{
get_data(list, path, type, RData = FALSE)
}
\arguments{
\item{list}{A list of lists. Each inner list contains the name of the file
where the remperature and discharge data, or the meteoSwiss data, are located.
If data to load are temperature and discharge data, a vector containing the
three capital letters apreviation of the associated meteoSwiss stations must
also be in the inner lists for each station.}

\item{path}{The relative path the prepend to the files to be read.}

\item{type}{Either \code{"WATER"} or \code{"METEO"} to specify if the data to be
loaded are meteoSwiss data}

\item{RData}{A \code{RDS} objec tto be loaded. If provided data won't be loaded
from SMET files}
}
\value{
A list of SMET object continning the data. The keys are the input files
names (without extension).
}
\description{
Read raw of river or meteo data. Data can be read from files or from RDS data.
All files to be read must be SMET files (see
\url{https://models.slf.ch/docserver/meteoio/SMET_specifications.pdf}).
Returns a list of SMET_OBJECT (see \code{\link{get_file_data}}), each entry
corresponding to one station.
}
