% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_day}
\alias{get_day}
\title{Returns day from a PosixCT (or vector of PosixCT)}
\usage{
get_day(timestamp)
}
\arguments{
\item{timestamp}{A PosixCT or vector of PosixCT}
}
\value{
The corresponding day (as numeric)
}
\description{
Returns day from a PosixCT (or vector of PosixCT)
}
