% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_preprocessing.R
\name{get_file_data}
\alias{get_file_data}
\title{Read data from a SMET file, returns a SMET_OBJECT}
\usage{
get_file_data(file, start = NULL, end = NULL, meteo = NULL)
}
\arguments{
\item{file}{Path of the SMET file to be loaded}

\item{start}{Starting year to keep (default \code{NULL})}

\item{end}{Ending year to keep(default \code{NULL})}

\item{meteo}{Three capital letters apreviation of the associated meteoSwiss
stations (default \code{NULL})}
}
\value{
SMET_OBJECT
}
\description{
Read data from a SMET file, returns a SMET_OBJECT with only full years,
bisssextile days removed, daily, monthly, seasonal, and yearly means computed
optionally cut between indicated years, and optoptionally associated with a list
of meteo stations.
}
\details{
The inner data are \code{$header} (SMET header),\code{$data} (raw data),
\code{$daily} (daily means), \code{$monthly} (monthly means), \code{$yearly} (yearly means),
\code{$DJF} (winter means), \code{$MAM} (spring means),
\code{$JJA} (summer means), and\code{$SON} (fall means). The meteo station
list is accissible through \code{$meteo}.
}
