% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_preprocessing.R
\name{get_file_data_only}
\alias{get_file_data_only}
\title{Read data from a SMET file, returns a SMET_OBJECT without means computed}
\usage{
get_file_data_only(file, start = NULL, end = NULL, meteo = NULL)
}
\arguments{
\item{file}{Path of the SMET file to be loaded}

\item{start}{Starting year to keep (default \code{NULL})}

\item{end}{Ending year to keep (default \code{NULL})}

\item{meteo}{Three capital letters apreviation of the associated meteoSwiss
stations (default \code{NULL})}
}
\value{
A list of SMET object continning the data. The keys are the input files
names (without extension).
}
\description{
Same as \code{\link{get_file_data}} but does not add the mean values.
}
