% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hysteresis_analysis.R
\name{get_hysteresis_data}
\alias{get_hysteresis_data}
\title{Produce hysteresis data}
\usage{
get_hysteresis_data(rivers_data, period, smoothing)
}
\arguments{
\item{rivers_data}{A list of SMET_OBJECT, which are obtained through the function
\code{\link{get_file_data}} and contain the data of one river station.}

\item{period}{The lenght (in year) of the periods over which the hystheresis
data should be computed}

\item{smoothing}{The length (in day) of the moving average window to
be applied}
}
\value{
The input SMET_OBJECT whih a new list entry, "hysteresis", containing
the hysteresis data. The hysteresyis data are discharge and temperature values,
averaged for each day of the year separately over various periods and smoothed
with a circular moving average window. The new "hysteresis" antry of the
SMET_OBJECT contains a list where the keys are the periods (in the format
"AAAA-BBBB" where AAAA is the starting year and BBBB the ending year) and the
associated data are a \code{\link[base]{data.frame}} containing discharge and
temperature values.
}
\description{
Read data from a list of SMET_OBJECT, which are obtained through the function
\code{\link{get_file_data}} and contain the data of one river station. It keeps
only subsequent years and compute necessary values for hysteresis plots
(daily means over years and smoothed daily means over years)
}
