% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_lm_summary}
\alias{get_lm_summary}
\title{Return the summary of \code{\link[stats]{lm}} model in a list of numeric values}
\usage{
get_lm_summary(lm)
}
\arguments{
\item{lm}{An object generated by \code{\link[stats]{lm}}}
}
\value{
A list with the folowing entries, all as numeric:
\item{intercept}{The intercept value}
\item{trend}{The trend value}
\item{intercept_std}{The std error of the intercept value}
\item{trend_std}{The std error of the trend value}
\item{intercept_p}{The p-value of the intercept value}
\item{trend_p}{The p-value of the trend value}
\item{r_squared}{The \eqn{R^{2}} value}
\item{adj_r_squared}{Teh adjusted R \eqn{R^{2}} value}
}
\description{
The values retuned are obtained by using \code{\link[base]{summary}}
on the \code{\link[stats]{lm}} object
}
