% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_month}
\alias{get_month}
\title{Returns month from a PosixCT (or vector of PosixCT)}
\usage{
get_month(timestamp)
}
\arguments{
\item{timestamp}{A PosixCT or vector of PosixCT}
}
\value{
The corresponding month (as numeric)
}
\description{
Returns month from a PosixCT (or vector of PosixCT)
}
