% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STL_analysis.R
\name{get_STL_analysis}
\alias{get_STL_analysis}
\title{Genenral call to perform STL analysis.}
\usage{
get_STL_analysis(rivers_data, meteo_data, variable, s_win, frequency,
  start = NULL, end = NULL)
}
\arguments{
\item{rivers_data}{A SMET_OBJECT with river data}

\item{meteo_data}{A SMET_OBJECT with meteo data}

\item{variable}{Variable on which the STL analysis should be performed}

\item{s_win}{Vector of \code{s.window} values for the stl analysis
(see \code{link[stats]{stl}})}

\item{frequency}{Frequency of data do be used. "monthly" or "monthly".}

\item{start}{Year to start the STL analysis (years before are cut).
Default \code{NULL}.}

\item{end}{Year to end the STL analysis (years after are cut).
Default \code{NULL}.}
}
\value{
A list where each key the station name and each value is a list.
In this list each key is a value of \code{s.window}. The content in
each list entry is the output of the STL analysis. In addiditon, the entry
\code{$meteo} gives access to a list where keys are the meteo station three
capital letters abbreviation and the content is a list with keys \code{s.window}
containing the STL analysis for the meteorological variable.
}
\description{
Genenral call to perform STL analysis.
}
