% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STL_analysis.R
\name{get_STL_output}
\alias{get_STL_output}
\title{Perform STL analysis}
\usage{
get_STL_output(station_data, meteo_data, variable, s_win, frequency)
}
\arguments{
\item{station_data}{A SMET_OBJECT with river data}

\item{meteo_data}{A SMET_OBJECT with meteo data}

\item{variable}{Variable on which the STL analysis should be performed}

\item{s_win}{vector of \code{s.window} values for the stl analysis
(see \code{link[stats]{stl}})}

\item{frequency}{Frequency of data do be used. "monthly" or "monthly".}
}
\value{
A list where each key is a value of \code{s.window}. The content in
each list entry is the output of the STL analysis. In addiditon, the entry
\code{$meteo} gives access to a list where keys are the meteo station three
capital letters abbreviation and the content is a list with keys \code{s.window}
containing the STL analysis for the meteorological variable.
}
\description{
Perform STL analysis for the given variable and the associated mete variable
(air temperature if the variable is water temperature and precipitation if
the river variable is discharge)
}
