% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STL_analysis.R
\name{get_time_series}
\alias{get_time_series}
\title{Convert data from SMET_OBJECT to ts objects}
\usage{
get_time_series(data)
}
\arguments{
\item{data}{A SMET_OBJECT}
}
\value{
A SMET_OBJECT with all inner \code{\link[base]{data.frame}} transformed
into lists of \code{\link[stats]{ts}} objects
}
\description{
Read a SMET_OBJECT, returns a list of \code{\link[base]{data.frame}} containing
data as \code{\link[stats]{ts}} objects. Input data.frames are split
into list of times series, each list entry being one variable.
}
