% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_preprocessing.R
\name{get_timestep}
\alias{get_timestep}
\title{Returns the timestep of the timeseries in a SMET_OBJECT}
\usage{
get_timestep(data)
}
\arguments{
\item{data}{SMET_OBJECT}
}
\value{
Timestep of the timeseries
}
\description{
Returns the timestep of the timeseries in a SMET_OBJECT. An error is thrown
if the timestep is not constant
}
