% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_year}
\alias{get_year}
\title{Returns year from a PosixCT (or vector of PosixCT)}
\usage{
get_year(timestamp)
}
\arguments{
\item{timestamp}{A PosixCT or vector of PosixCT}
}
\value{
The corresponding year (as numeric)
}
\description{
Returns year from a PosixCT (or vector of PosixCT)
}
