% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_preprocessing.R
\name{keep_subs_years_preprocessing}
\alias{keep_subs_years_preprocessing}
\title{Filter tiem series for complete years}
\usage{
keep_subs_years_preprocessing(data)
}
\arguments{
\item{data}{A SMET_OBJECT}
}
\value{
A SMET_OBJECT
}
\description{
Remove years starting or ending with NaN, kepps only subsequent years,
starting from the end of the timeseries.
}
