% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_functions.R
\name{plot_acf}
\alias{plot_acf}
\title{Plot acf and pacf of the residuals of the STL analysis}
\usage{
plot_acf(station, output_type = "NONE")
}
\arguments{
\item{station}{A SMET object containing the data for one station}

\item{output_type}{Either \code{"NONE"} (default), \code{"PDF"} or \code{"PNG"}.

\code{output_type = "NONE"} creates the plot in a normal plot window (default),

\code{output_type = "PDF"} saves the plot as pdf in plots/analysis/'station_name'/,

\code{output_type = "PNG"} saves the plot as png in plots/analysis/'station_name'/,}
}
\description{
This function plots the acf and pacf of the residuals of the STL analysis
for the four variables T, Q, TA and P for the water station passed in
parameters and the associated meteoSwiss stations. This produces the plots
shown in Figures S7 and S8 in supplementary.
}
