% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_alpine.R
\name{plot_alpine}
\alias{plot_alpine}
\title{Produces plot for alpine catchemts}
\usage{
plot_alpine(rivers_data)
}
\arguments{
\item{rivers_data}{The dataset of rivers data}
}
\description{
This function produces plot for alpine catchments (Figure S35). The plots are
saved in plots/alpine.pdf
}
