% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_correlation_plots.R
\name{plot_correlations}
\alias{plot_correlations}
\title{Print correlation matrices}
\usage{
plot_correlations(rivers_data)
}
\arguments{
\item{rivers_data}{The dataset of rivers data}
}
\description{
This function prints to the console the correlations matrices shown in
Tables 3 and in Tables S6 and S7 in supplementary.
}
\details{
Some additional plots nor present in the paper are also produced and
saveud under plots/correlations_plots.pdf
}
