% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{plot_general_situation}
\alias{plot_general_situation}
\title{Plot a map with the location of the river station and meteostation used}
\usage{
plot_general_situation(rivers_data, output_type = "NONE")
}
\arguments{
\item{rivers_data}{A list of SMET_OBJECT obtained through the function
\code{\link{get_file_data}} containing
the data on the rivers stations}

\item{output_type}{Either \code{"NONE"} (default), \code{"PDF"} or \code{"PNG"}.

\code{output_type = "NONE"} creates the plot in a normal plot window,

\code{output_type = "PDF"} saves the plot as pdf under plots/General_situation.pdf,

\code{output_type = "PNG"} saves the plot as png under plots/General_situation.png}
}
\description{
Plot a map with the location of the river station and meteostation used
}
\section{Requirements}{

This functions needs the following files to be abailable:
maps/processed_maps/swiss_map.tif, maps/processed_maps/lakes.shp,
maps/processed_maps/rivers.shp, maps/processed_maps/borders.shp,
meteo/MeteoSwiss_StationList.txt. In addition, the plot directory must exist.
}

