% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_functions.R
\name{plot_hysteresis}
\alias{plot_hysteresis}
\title{Plot hysteresis}
\usage{
plot_hysteresis(station, output_type = "NONE")
}
\arguments{
\item{station}{A SMET object containing the data for one station}

\item{output_type}{Either \code{"NONE"} (default), \code{"PDF"} or \code{"PNG"}.

\code{output_type = "NONE"} creates the plot in a normal plot window (default),

\code{output_type = "PDF"} saves the plot as pdf in plots/analysis/'station_name'/,

\code{output_type = "PNG"} saves the plot as png in plots/analysis/'station_name'/,}
}
\description{
This function plots the day-of-the -year decadal mean of Q and T for the
given station along wiht the Q-T hysteresis plot. This function produces the
plot shown in Figure 15
}
