% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_lakes.R
\name{plot_lakes}
\alias{plot_lakes}
\title{Produces plots for lakes}
\usage{
plot_lakes(rivers_data)
}
\arguments{
\item{rivers_data}{The dataset of rivers data}
}
\description{
This function produces the plots for the trends before and after lakes, shown
in Figure 7 and in Figures S14 to S17 in supplementary. Figures are saved
under plots/lakes_plots.pdf
}
