% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STL_analysis.R
\name{plot_remainder_analysis}
\alias{plot_remainder_analysis}
\title{Title}
\usage{
plot_remainder_analysis(STL_output, s_windows, version, name)
}
\arguments{
\item{STL_output}{a}

\item{s_windows}{a}

\item{version}{a}

\item{name}{a}
}
\value{
a
}
\description{
Title
}
