% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_snow.R
\name{plot_snow}
\alias{plot_snow}
\title{Produce snow cover and glacier mass balance plots}
\usage{
plot_snow(rivers_data)
}
\arguments{
\item{rivers_data}{The dataset of rivers data}
}
\description{
This function produces the monthly snow cover plots (Figures 12 and S21 in
suplpementary) and the glaciers mass balance plot (Figure S23 in supplementary).
The plots are saved in plots/snow_plots.pdf
}
