% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_functions.R
\name{plot_stl}
\alias{plot_stl}
\title{Plot component of STL for one station}
\usage{
plot_stl(station, output_type = "NONE")
}
\arguments{
\item{station}{A SMET object containing the data for one station}

\item{output_type}{Either \code{"NONE"} (default), \code{"PDF"} or \code{"PNG"}.

\code{output_type = "NONE"} creates the plot in a normal plot window,

\code{output_type = "PDF"} saves the plot as pdf under plots/General_situation.pdf,

\code{output_type = "PNG"} saves the plot as png under plots/General_situation.png}
}
\description{
This function plots the component of the STL for the four variables T, Q, TA
and P for the water station passed in parameters and the associated meteoSwiss
stations. This produces the plots shown in Figures S1 to S4 in supplementary.
}
