% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_thresholds.R
\name{plot_thresholds}
\alias{plot_thresholds}
\title{Returns plots for the 15C and 25C thresholds analysis}
\usage{
plot_thresholds(rivers_data, rivers_data_1h)
}
\arguments{
\item{rivers_data}{The dataset of rivers data}

\item{rivers_data_1h}{The dataset of rivers data at 1 hour resolution}
}
\description{
This function return plots for the 15C and 25C thresholds analysis shown
in Figures 17 and 18 and in Figure S24. PLots are saved under
plots/25_degs.pdf and plots/15_degs.pdf
}
