% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_functions.R
\name{plot_trends}
\alias{plot_trends}
\title{Plot time series and trends}
\usage{
plot_trends(station, output_type = "NONE")
}
\arguments{
\item{station}{A SMET object containing the data for one station}

\item{output_type}{Either \code{"NONE"} (default), \code{"PDF"} or \code{"PNG"}.

\code{output_type = "NONE"} creates the plot in a normal plot window (default),

\code{output_type = "PDF"} saves the plot as pdf in plots/analysis/'station_name'/,

\code{output_type = "PNG"} saves the plot as png in plots/analysis/'station_name'/,}
}
\description{
This function produces plots for time series and trends for the given station
and for the variables T, Q, TA, P. These plots are not used in the paper.
}
