% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_yearly_anomalies.R
\name{plot_yearly_anomalies}
\alias{plot_yearly_anomalies}
\title{Plot yearyl seasonnal anomlaies}
\usage{
plot_yearly_anomalies(rivers_data)
}
\arguments{
\item{rivers_data}{The dataset of rivers data}
}
\description{
This function plots the yearly seasonnal anomalies for T, Q, TA and P shown
in Figures 11 and 13 and in Figures S20 and S22 in supplementary. The figure
are saved under plots/summer_anomalies.pdf, plots/winter_anomalies.pdf,
plots/spring_anomalies.pdf, and plots/fall_anomalies.pdf
}
