% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STL_analysis.R
\name{post_season_loess}
\alias{post_season_loess}
\title{Do an additional Loess fitting on the STL analysis}
\usage{
post_season_loess(stl_input, s_win)
}
\arguments{
\item{stl_input}{List of stl analysis output, keys being the used \code{s.window}
values}

\item{s_win}{Vector of \code{s.window} values for the stl analysis
(see \code{link[stats]{stl}})}
}
\value{
List of stl analysis output, keys being the used \code{s.window}
values
}
\description{
Do an additional Loess fitting on the STL analysis seasonnal signal as suggested
in R. B. Cleveland, W. S. Cleveland, J.E. McRae, and I. Terpenning (1990)
STL: A Seasonal-Trend Decomposition Procedure Based on Loess. Journal
of Official Statistics, 6, 3–73
}
