% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_print_all_trends.R
\name{print_all_trends}
\alias{print_all_trends}
\title{Print trends table in Latex format}
\usage{
print_all_trends(rivers_data, meteo_data)
}
\arguments{
\item{rivers_data}{The dataset of rivers data}

\item{meteo_data}{The dataset of homegenous MeteoSwiss data}
}
\description{
Tihs function prints to the console the trends table shown in Tables A1 and
A2 in appendix and in Tables S1 and S2 in supplementary. The table are
printed in latex format
}
