% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_preprocessing.R
\name{read_smet}
\alias{read_smet}
\title{Read SMET file, returns a SMET object (a list)}
\usage{
read_smet(file_name)
}
\arguments{
\item{file_name}{String containing the path th file to read}
}
\value{
A SMET_OBJECT containing the data
}
\description{
Read SMET file, returns a SMET object (a list). The SMET_OBJECT has
the following structure:
\describe{
\item{\code{$header}}{The header information, based on smet header value}
\item{\code{$data}}{A \code{\link[base]{data.frame}} containing read data, column names from SMET header,
timestamp as PosixCT}
}
}
