% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input_preprocessing.R
\name{remove_bissextile}
\alias{remove_bissextile}
\title{Remove 29th of february from a SMET_OBJECT}
\usage{
remove_bissextile(data)
}
\arguments{
\item{data}{SMET_OBJECT}
}
\value{
SMET_OBJECT with 29th of February removed
}
\description{
Remove 29th of february from a SMET_OBJECT
}
