% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hysteresis_analysis.R
\name{smooth_circular}
\alias{smooth_circular}
\title{Compute a circular moving window average over all columns of a
\code{\link[base]{data.frame}} containing daily data over a year (365 data)}
\usage{
smooth_circular(year_daily_means, smooth_time)
}
\arguments{
\item{year_daily_means}{A \code{\link[base]{data.frame}} with numeric values
containing daily data over a year (365 data)}

\item{smooth_time}{The window to be used for the moving average}
}
\value{
The \code{\link[base]{data.frame}} smoothed
}
\description{
Compute a circular moving window average over all columns of a
\code{\link[base]{data.frame}} containing daily data over a year (365 data)
}
