% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hysteresis_analysis.R
\name{smooth_daily_means_over_period}
\alias{smooth_daily_means_over_period}
\title{Function wrapper for \code{\link{smooth_circular}}}
\usage{
smooth_daily_means_over_period(daily_means, smooth_time)
}
\arguments{
\item{daily_means}{A list of \code{\link[base]{data.frame}}.
Each \code{\link[base]{data.frame}} is made of numeric values and
contains daily data over a year (365 data)}

\item{smooth_time}{The window to be used for the moving average}
}
\value{
A list where each key is one time period. List emntries are
\code{\link[base]{data.frame}} of daily mean data circularly smoothed
over the period used as key. Periods have the format "AAAA-BBBB" where
AAAA is the starting year and BBBB the ending year
}
\description{
Wrapper for \code{\link{smooth_circular}} to call it over all the periods
defined in \code{\link{compute_daily_means_over_period}}
}
