% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{trim}
\alias{trim}
\title{Trim spaces in string}
\source{
Function taken from \url{https://stackoverflow.com/questions/2261079/how-to-trim-leading-and-trailing-whitespace}
}
\usage{
trim(x)
}
\arguments{
\item{x}{A string}
}
\value{
Input string with leading or trailing spaces removed
}
\description{
Trim spaces in string
}
